+++ examples/DashboardBot/Program.cs
@@ -26,8 +26,7 @@ public class Program
             Intents = PawSharp.Core.Enums.GatewayIntents.AllNonPrivileged
         };
 
-        services.AddSingleton(options);
-        services.AddPawSharpClient();
+        services.SetupPawSharp(options);
         services.AddPawSharpCommands();
         services.AddPawSharpInteractions();
 
@@ -123,7 +122,7 @@ public class DashboardCommands : InteractionModule
             .AddField("Servers", client.Guilds.Count.ToString(), true)
             .AddField("Users", client.Guilds.Sum(g => g.MemberCount).ToString(), true)
             .AddField("Uptime", "Running", true)
-            .AddField("Version", "PawSharp 0.5.0-alpha10", false)
+            .AddField("Version", "PawSharp 1.0.0-alpha.2", false)
             .WithColor(Color.Purple);
 
         await RespondAsync(embed: embed.Build());